+++ docs/release-notes.md
@@ -16,6 +16,7 @@ nav_order: 9
 ### Bug fixes
 
 - Fail if LUKS method is not specified while `boot_device.luks.device` is set _(fcos 1.7.0-exp)_
+- TODO: Add Kernel Argument rd.luks.keys in s390x layout sugar.
 
 ### Misc. changes